renamed pathresolver to pathparser for a bit more clarity
