+++ pkg/pathparser/pathparser.go
@@ -1,13 +1,14 @@
-package pathresolver
+package pathparser
 
 import (
+	"fmt"
 	"strings"
 
 	"github.com/matthewchivers/rb/pkg/fsutil"
 )
 
-// ExpandPath expands a path to an absolute path
-func ExpandPath(fs fsutil.FSUtil, path string) (string, error) {
+// ParsePath checks for validity and expands a path
+func ParsePath(fs fsutil.FSUtil, path string) (string, error) {
 	if strings.HasPrefix(path, "~") {
 		return expandTilde(fs, path)
 	}
@@ -20,7 +21,7 @@ func ExpandPath(fs fsutil.FSUtil, path string) (string, error) {
 func expandTilde(fs fsutil.FSUtil, path string) (string, error) {
 	home, err := fs.UserHomeDir()
 	if err != nil {
-		return path, err
+		return path, fmt.Errorf("could not expand tilde path: %v", err)
 	}
 	return strings.Replace(path, "~", home, 1), nil
 }
@@ -28,7 +29,7 @@ func expandTilde(fs fsutil.FSUtil, path string) (string, error) {
 func expandRelativePath(fs fsutil.FSUtil, path string) (string, error) {
 	absPath, err := fs.Abs(path)
 	if err != nil {
-		return path, err
+		return path, fmt.Errorf("could not expand relative path: %v", err)
 	}
 	return absPath, nil
 }
